all modals working + new structure
