+++ midas-portal/src/app/components/landing/landing.module.ts
@@ -5,11 +5,10 @@ import { LandingComponent } from './landing.component';
 import {HttpClientModule}   from '@angular/common/http';
 import { HttpClient } from '@angular/common/http';
 import { AngularMaterialModule } from '../../angular-material.module';
-import { EdiListComponent } from '../edi-list/edi-list.component';
-import { RecordsComponent } from '../records/records.component';
-import { DmpListComponent } from '../dmp-list/dmp-list.component';
-import { ReviewListComponent } from '../review-list/review-list.component';
-import { FileListComponent } from '../file-list/file-list.component';
+import { DapComponent } from '../tables/dap/dap.component';
+import { DmpListComponent } from '../tables/dmp-list/dmp-list.component';
+import { ReviewListComponent } from '../tables/review-list/review-list.component';
+import { FileListComponent } from '../tables/file-list/file-list.component';
 import { ToastModule } from 'primeng/toast';
 import { BrowserModule } from '@angular/platform-browser';
 import { PanelModule } from "primeng/panel";
@@ -24,17 +23,22 @@ import {TableModule} from 'primeng/table';
 import { DropdownModule } from 'primeng/dropdown';
 import { FormsModule } from '@angular/forms';
 import { DatePipe } from '@angular/common';
-import { RecordsExtComponent } from '../ext/records/records-ext.component';
+import { DapModalComponent } from '../modals/dap/dap.component';
+import { FileListModalComponent } from '../modals/file-list/file-list.component';
+import { DmpListModalComponent } from '../modals/dmp-list/dmp-list.component';
+import { ReviewListModalComponent } from '../modals/review-list/review-list.component';
 
 @NgModule({
     declarations:[
         LandingComponent,
-        EdiListComponent,
-        RecordsComponent,
+        DapComponent,
         DmpListComponent,
         ReviewListComponent,
         FileListComponent,
-        RecordsExtComponent
+        DapModalComponent,
+        FileListModalComponent,
+        DmpListModalComponent,
+        ReviewListModalComponent
     ],
     imports:[
          HttpClientModule,
