+++ midas-portal/src/app/components/modals/file-list/file-list.component.html
@@ -4,9 +4,10 @@
             <!-- <span class="pi pi-cog"></span> -->
             <a href={{nextcloudUI}} target="_blank"><fa-icon [style]="{display: 'inline-block', padding: '5px', color: 'grey'}" [icon]="faFileImport" size="lg" title="Add or update files"></fa-icon></a>
         </button>
-
+        <p-toast></p-toast>
         <!-- <p-menu #menu id="config_menu" [model]="items" [popup]="true"></p-menu> -->
 </ng-template>
+    
     <p-table #filetable [value]="data" [paginator]="true"
         [rows]="5" [scrollable]="true"
         [showCurrentPageReport]="true"
@@ -41,4 +42,5 @@
         </ng-template>
     </p-table>
 </p-panel>
+
