+++ midas-portal/src/app/components/tables/dap/dap.component.html
@@ -5,7 +5,7 @@
             <a href={{dapUI}} target="_blank"><fa-icon [style]="{display: 'inline-block', padding: '5px', color: 'grey'}" [icon]="faFileEdit" size="lg" title="Create New Record"></fa-icon></a> 
         </button>
         <p-toast></p-toast>
-<button pButton class="p-panel-header-icon p-link" (click)="show()"><fa-icon [style]="{display: 'inline-block', color: '#005EA2'}" [icon]="faUpRightAndDownLeftFromCenter" size="lg" title="Create New Record"></fa-icon></button>
+<button pButton class="p-panel-header-icon p-link" (click)="show()"><fa-icon [style]="{display: 'inline-block', color: '#005EA2'}" [icon]="faUpRightAndDownLeftFromCenter" size="lg" title="Expand"></fa-icon></button>
     </ng-template>
         <p-table #recordsTable name="recordtable" [value]="data" [paginator]="true"
             height="100px" [rows]="4" [scrollable]="true"
