+++ midas-portal/src/app/components/tables/dmp-list/dmp-list.component.html
@@ -1,0 +1,42 @@
+
+<div class="dmp">
+<p-panel header="My Data Management Plans">
+    <ng-template pTemplate="icons">
+        <button pButton class="p-panel-header-icon p-link" (click)="titleClick()" >
+            <fa-icon [style]="{display: 'inline-block', padding: '5px', color: 'green'}" [icon]="faCheck" size="lg" title="All DMPs completed."></fa-icon> 
+        </button>
+        <button pButton class="p-panel-header-icon p-link" (click)="titleClick()">
+            <a href={{dmpUI}} target="blank"><fa-icon [style]="{display: 'inline-block', padding: '5px', color: 'grey'}" [icon]="faFileEdit" size="lg" title="Create New Data management plan"></fa-icon> </a>
+        </button>
+        <p-toast></p-toast>
+<button pButton class="p-panel-header-icon p-link" (click)="show()"><fa-icon [style]="{display: 'inline-block', color: '#005EA2'}" [icon]="faUpRightAndDownLeftFromCenter" size="lg" title="Expand"></fa-icon></button>
+        <!-- <p-menu #menu id="config_menu" [model]="items" [popup]="true"></p-menu> -->
+</ng-template>
+    <p-table #dmptable [value]="data" [paginator]="true" 
+        height="100px" [rows]="5" [scrollable]="true"
+        [showCurrentPageReport] = "true"
+        [rowHover]="true"
+        currentPageReportTemplate="Showing {first} to {last} of {totalRecords} entries" 
+        [rowsPerPageOptions]="[5]" [globalFilterFields]="['name', 'owner','nist']">
+        <ng-template pTemplate="header">
+            <tr>
+                <th pResizableColumn pSortableColumn="name">Name<p-sortIcon field="name"></p-sortIcon></th>
+                <th pResizableColumn pSortableColumn="owner">Owner<p-sortIcon field="owner"></p-sortIcon></th>
+                <th pResizableColumn pSortableColumn="nist">Primary NIST Contact<p-sortIcon field="nist"></p-sortIcon></th>
+                <th pResizableColumn>Last Modified
+                    <p-columnFilter title="Filter by date" type="date" field="status.modifiedDate" display="menu"></p-columnFilter>
+                </th>
+            </tr>
+        </ng-template>
+        <ng-template pTemplate="body" let-item>
+            <tr>
+                <td>{{item.name}}</td>
+                <td>{{item.owner}}</td>
+                <td>{{item.data.primary_NIST_contact.firstName+' '+item.data.primary_NIST_contact.lastName}}</td>
+                <td>{{this.datepipe.transform(item.status.modifiedDate,'MM/dd/yyyy')}}</td>
+            </tr>
+        </ng-template>
+    </p-table>
+</p-panel>
+    
+</div>
