+++ midas-portal/src/app/components/tables/file-list/file-list.component.css
@@ -1,0 +1,6 @@
+.td-class {
+  word-wrap: break-word;
+  word-break: break-all;
+  white-space: normal;
+  padding: 0.2rem;
+}
