+++ midas-portal/src/app/components/tables/file-list/file-list.component.html
@@ -1,0 +1,47 @@
+<div *ngIf="display==true; else noData ">
+<p-panel header="My Files">
+    <ng-template pTemplate="icons">
+        <button pButton class="p-panel-header-icon p-link" (click)="titleClick()">
+            <!-- <span class="pi pi-cog"></span> -->
+            <a href={{nextcloudUI}} target="_blank"><fa-icon [style]="{display: 'inline-block', padding: '5px', color: 'grey'}" [icon]="faFileImport" size="lg" title="Add or update files"></fa-icon></a>
+        </button>
+        <p-toast></p-toast>
+<button pButton class="p-panel-header-icon p-link" (click)="show()"><fa-icon [style]="{display: 'inline-block', color: '#005EA2'}" [icon]="faUpRightAndDownLeftFromCenter" size="lg" title="Expand"></fa-icon></button>
+        <!-- <p-menu #menu id="config_menu" [model]="items" [popup]="true"></p-menu> -->
+</ng-template>
+    
+    <p-table #filetable [value]="data" [paginator]="true"
+        [rows]="5" [scrollable]="true"
+        [showCurrentPageReport]="true"
+        [rowHover]="true"
+        currentPageReportTemplate="Showing {first} to {last} of {totalRecords} entries" 
+        [rowsPerPageOptions]="[5]" [globalFilterFields]="['file_name', 'file_path']">
+        <ng-template pTemplate="header">
+            <tr>
+                <th pResizableColumn pSortableColumn="file_name">File Name <p-sortIcon field="file_name"></p-sortIcon></th>
+                <th pResizableColumn>File Path</th>
+                <th pResizableColumn pSortableColumn="file_size">File Size</th>
+                <th pResizableColumn pSortableColumn="last_modified"> Last Modified
+                    <p-columnFilter title="Filter by date" type="date" field="last_modified" display="menu"></p-columnFilter>
+                </th>
+            </tr>
+        </ng-template>
+        <ng-template pTemplate="body" let-item>
+            <tr>
+                <td class="td-class">{{item.file_name}}</td>
+                <td class="td-class">{{item.file_path}}</td>
+                <td class="td-class">{{item.file_size}}</td>
+                <td>{{this.datepipe.transform(item.last_modified,'MM/dd/yyyy')}}</td>
+            </tr>
+        </ng-template>
+    </p-table>
+</p-panel>
+</div>
+<ng-template #noData>
+    <p-panel header="My Files">
+        No data to display for the user cnd7
+    </p-panel>
+    
+</ng-template>
+
+
