+++ midas-portal/src/app/components/tables/file-list/file-list.component.ts
@@ -1,11 +1,14 @@
 import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import {faFileImport} from '@fortawesome/free-solid-svg-icons';
+import {faFileImport,faUpRightAndDownLeftFromCenter} from '@fortawesome/free-solid-svg-icons';
 import { Table } from 'primeng/table';
 import { AppConfig } from 'src/app/config/app.config';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 import { DatePipe } from '@angular/common';
-
+import { DialogService,DynamicDialogRef } from 'primeng/dynamicdialog';
+import { MessageService } from 'primeng/api';
+import { MatDialogRef } from '@angular/material/dialog';
+import { FileListModalComponent } from '../../modals/file-list/file-list.component';
 
 @Component({
   selector: 'app-file-list',
@@ -13,43 +16,64 @@ import { DatePipe } from '@angular/common';
   styleUrls: ['./file-list.component.css']
 })
 export class FileListComponent implements OnInit {
-
+  faUpRightAndDownLeftFromCenter=faUpRightAndDownLeftFromCenter;
   faFileImport=faFileImport;
   public records: any;
   public recordsApi: string;
   public data: any;
   public nextcloudUI: string;
+  public display:boolean;
+  ref: DynamicDialogRef;
   
   
 
   @ViewChild('filetable') fileTable: Table;
 
-  constructor(private appConfig: AppConfig,private http:HttpClient,public datepipe:DatePipe) { 
+  constructor(private appConfig: AppConfig,private http:HttpClient,public datepipe:DatePipe,public dialogService: DialogService
+    , public messageService: MessageService) { 
     this.recordsApi = 'https://data.nist.gov/rmm/records'
     
   }
+
+
 
   
   
 
   async ngOnInit() {
+    this.display=true;
     let promise = new Promise((resolve) => {
       this.appConfig.getRemoteConfig().subscribe(config => {
         this.nextcloudUI = config.nextcloudUI;
         resolve(this.nextcloudUI);
         //GET method to get data
         this.fetchRecords(this.nextcloudUI);
+        console.log(this.data)
+        if(this.data==null){
+          this.display=false;
+        }else{
         for (let i = 0; i<this.data.length;i++){
-          this.data[i].status.modifiedDate = new Date(this.data[i].status.modifiedDate)
+          this.data[i].last_modified = new Date(this.data[i].last_modified)
         }
+      }
+        
       });
+      
     });
     
     //await this.getRecords()
     //this.data = this.records.ResultData
     //this.data = RECORD_DATA;
-    console.log(this.data)
     
+  }
+
+  show() {
+    this.ref = this.dialogService.open(FileListModalComponent, {
+        data:this.data,
+        width: '80%',
+        contentStyle: { overflow: 'auto' },
+        baseZIndex: 10000,
+    });
   }
 
   async getRecords() {
