+++ midas-portal/src/app/components/tables/review-list/review-list.component.html
@@ -1,0 +1,43 @@
+
+<p-panel header="My Reviews">
+    <ng-template pTemplate="icons">
+        <button pButton class="p-panel-header-icon p-link" (click)="titleClick()">
+            <!-- <span class="pi pi-cog"></span> -->
+            <fa-icon [style]="{display: 'inline-block', padding: '5px', color: 'orange'}" [icon]="faBell" size="lg" title="You have new Records to review."></fa-icon>
+       
+        </button>
+        <button pButton class="p-panel-header-icon p-link" (click)="titleClick()">
+            <!-- <span class="pi pi-cog"></span> -->
+            <a href={{npsUI}} target="_blank"><fa-icon [style]="{display: 'inline-block', padding: '5px', color: 'grey'}" [icon]="faListCheck" size="lg" title="Check Reviews"></fa-icon></a>
+       
+        </button>
+        <p-toast></p-toast>
+<button pButton class="p-panel-header-icon p-link" (click)="show()"><fa-icon [style]="{display: 'inline-block', color: '#005EA2'}" [icon]="faUpRightAndDownLeftFromCenter" size="lg" title="Expand"></fa-icon></button>
+        
+        <!-- <p-menu #menu id="config_menu" [model]="items" [popup]="true"></p-menu> -->
+</ng-template>
+<div class="">
+    <p-table #reviewtable [loading]="loading" [rowHover]="true"[value]="data"
+        [paginator]="true" [rows]="5" [scrollable]="true"
+        [showCurrentPageReport]="true"
+        currentPageReportTemplate="Showing {first} to {last} of {totalRecords} entries" 
+        [rowsPerPageOptions]="[5]" [globalFilterFields]="['title', 'submittername', 'currentreviewer','reviewstep']">
+        <ng-template pTemplate="header">
+            <tr>
+                <th style="width:20%" pSortableColumn="title">Title <p-sortIcon field="title"></p-sortIcon></th>
+                <th style="width:20%" pSortableColumn="submittername">Submitter Name <p-sortIcon field="submittername"></p-sortIcon></th>
+                <th style="width:20%" pSortableColumn="currentreviewer">Current Reviewer <p-sortIcon field="currentreviewer"></p-sortIcon></th>
+                <th style="width:20%" pSortableColumn="reviewstep">Current Review Step <p-sortIcon field="reviewstep"></p-sortIcon></th>
+            </tr>
+        </ng-template>
+        <ng-template pTemplate="body" let-item>
+            <tr class="tr-class">
+                <td class="td-class">{{item.title}}</td>
+                <td class="td-class">{{item.submitterName}}</td>
+                <td class="td-class">{{item.currentReviewer}}</td>
+                <td class="td-class">{{item.currentReviewStep}}</td>
+            </tr>
+        </ng-template>
+    </p-table>
+</div>
+</p-panel>
